added key changer texture | fixed portable radar crash | fixed language strings regarding block options |

+++ 1.7.10/src/main/java/net/geforcemods/securitycraft/api/Option.java
@@ -212,7 +212,7 @@ public static class OptionInt extends Option<Integer>{
 	}
 	
 	public String toString() {
-		return ((int) value) + "";
+		return (value) + "";
 	}
 }
 
@@ -281,7 +281,7 @@ public static class OptionFloat extends Option<Float>{
 	}
 	
 	public Float getValue() {
-		return (Float) value;
+		return value;
 	}
 	
 	public String toString() {
