+++ 1.7.10/src/main/java/net/geforcemods/securitycraft/api/TileEntitySCTE.java
@@ -304,7 +304,7 @@ public class TileEntitySCTE extends TileEntity implements INameable{
      * <p>
      * Calls {@link TileEntitySCTE}.attackEntity(Entity) when this TE's cooldown equals 0.
      */ 
-    public TileEntitySCTE attacks(Class<?> type, int range, int cooldown) {
+    public TileEntitySCTE attacks(Class<?> type, double range, int cooldown) {
     	attacks = true;
     	typeToAttack = type;
     	attackRange = range;
